+++ packages/server/test/complete.test.ts
@@ -191,121 +191,6 @@ describe('on blank space', () => {
   })
 })
 
-describe('TableName completion', () => {
-  test("complete function keyword", () => {
-    const result = complete('SELECT arr', { line: 0, column: 10 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('array_concat()')
-    expect(result.candidates[1].label).toEqual('array_contains()')
-  })
-
-  test("complete function keyword", () => {
-    const result = complete('SELECT ARR', { line: 0, column: 10 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('ARRAY_CONCAT()')
-    expect(result.candidates[1].label).toEqual('ARRAY_CONTAINS()')
-  })
-
-  test("complete TableName", () => {
-    const result = complete('SELECT T FROM TABLE1', { line: 0, column: 8 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(1)
-    expect(result.candidates[0].label).toEqual('TABLE1')
-  })
-  test("complete alias", () => {
-    const result = complete('SELECT ta FROM TABLE1 as tab', { line: 0, column: 9 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(1)
-    expect(result.candidates[0].label).toEqual('tab')
-  })
-  test("complete SELECT star", () => {
-    const result = complete('SELECT FROM TABLE1', { line: 0, column: 6 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('Select all columns from TABLE1')
-    expect(result.candidates[0].insertText).toEqual('SELECT\nTABLE1.COLUMN1,\nTABLE1.COLUMN2')
-    expect(result.candidates[1].label).toEqual('SELECT')
-  })
-  // This is difficult because we can't parse the statement to get
-  // at the column names
-  test.skip("complete partial SELECT star", () => {
-    const result = complete('SELEC FROM TABLE1', { line: 0, column: 5 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('SELECT')
-    expect(result.candidates[1].label).toEqual('Select all columns from TABLE1')
-    expect(result.candidates[1].insertText).toEqual('SELECT\nTABLE1.COLUMN1,\nTABLE1.COLUMN2')
-  })
-  test("complete SELECT star passed select ", () => {
-    const result = complete('SELECT  FROM TABLE1', { line: 0, column: 7 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(13)
-    const expected = [
-      expect.objectContaining({ label: 'Select all columns from TABLE1', insertText: 'TABLE1.COLUMN1,\nTABLE1.COLUMN2' }),
-    ]
-    expect(result.candidates).toEqual(expect.arrayContaining(expected))
-  })
-})
-
-describe('ColumnName completion', () => {
-  test("complete column name", () => {
-    const result = complete('SELECT COL FROM TABLE1', { line: 0, column: 10 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(0)
-  })
-
-  test("complete ColumnName", () => {
-    const result = complete('SELECT TABLE1.C FROM TABLE1', { line: 0, column: 15 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    const expected = [
-      expect.objectContaining({ label: 'COLUMN1' }),
-      expect.objectContaining({ label: 'COLUMN2' }),
-    ]
-    expect(result.candidates).toEqual(expect.arrayContaining(expected))
-  })
-
-  test("complete ColumnName with previous back tick column", () => {
-    const result = complete('SELECT `COLUMN2`, TABLE1.C FROM TABLE1', { line: 0, column: 26 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('COLUMN1')
-    expect(result.candidates[1].label).toEqual('COLUMN2')
-  })
-
-  test("complete ColumnName: cursor on dot", () => {
-    const result = complete('SELECT TABLE1. FROM TABLE1', { line: 0, column: 14 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('COLUMN1')
-    expect(result.candidates[1].label).toEqual('COLUMN2')
-  })
-
-  test("complete ColumnName:cursor on dot:multi line", () => {
-    const result = complete('SELECT *\nFROM TABLE1\nWHERE TABLE1.', { line: 2, column: 13 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('COLUMN1')
-    expect(result.candidates[1].label).toEqual('COLUMN2')
-  })
-
-  test("complete ColumnName:cursor on dot:using alias", () => {
-    const result = complete('SELECT *\nFROM TABLE1 t\nWHERE t.', { line: 2, column: 8 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('COLUMN1')
-    expect(result.candidates[1].label).toEqual('COLUMN2')
-  })
-
-  test("complete ColumnName:cursor on dot:using alias", () => {
-    const result = complete('SELECT t. FROM TABLE1 as t', { line: 0, column: 9 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('COLUMN1')
-    expect(result.candidates[1].label).toEqual('COLUMN2')
-  })
-
-  test("complete ColumnName:cursor on first char:using alias", () => {
-    const result = complete('SELECT t.C FROM TABLE1 as t', { line: 0, column: 10 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(2)
-    expect(result.candidates[0].label).toEqual('COLUMN1')
-    expect(result.candidates[1].label).toEqual('COLUMN2')
-  })
-
-  test("complete ColumnName:cursor on first char:using back tick", () => {
-    const result = complete('SELECT `t`.C FROM TABLE1 as t', { line: 0, column: 12 }, SIMPLE_SCHEMA)
-    expect(result.candidates.length).toEqual(0)
-  })
-})
-
 describe('From clause', () => {
   test("from clause: complete TableName:single line", () => {
     const result = complete('SELECT TABLE1.COLUMN1 FROM T', { line: 0, column: 28 }, SIMPLE_SCHEMA)
